# backend/docs/swagger.yaml
definitions:
  delivery_http.HealthResponse:
    properties:
      status:
        type: string
    type: object
  delivery_http_user.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dto.CreateUserRequest:
    properties:
      city:
        example: Saint Petersburg
        maxLength: 50
        type: string
      fio:
        example: Петрова Мария Сергеевна
        maxLength: 255
        minLength: 1
        type: string
      status:
        enum:
        - active
        - blocked
        example: active
        type: string
      telephone_number:
        example: "+79998887766"
        maxLength: 20
        minLength: 5
        type: string
      user_description:
        example: Волонтёр приюта
        maxLength: 1000
        type: string
      user_login:
        example: maria_petrova
        maxLength: 50
        minLength: 3
        type: string
      user_password:
        example: securepassword123
        maxLength: 255
        minLength: 6
        type: string
    required:
    - fio
    - telephone_number
    - user_login
    - user_password
    type: object
  dto.LoginRequest:
    properties:
      user_login:
        example: maria_petrova
        maxLength: 50
        minLength: 3
        type: string
      user_password:
        example: securepassword123
        maxLength: 255
        minLength: 6
        type: string
    required:
    - user_login
    - user_password
    type: object
  dto.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiaXNfYWRtaW4iOmZhbHNlLCJleHAiOjE3MDAwMDAwMDB9
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.RegisterResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      city:
        example: Kazan
        maxLength: 50
        type: string
      fio:
        example: Петрова Мария Сергеевна
        maxLength: 255
        minLength: 1
        type: string
      status:
        enum:
        - active
        - blocked
        example: active
        type: string
      telephone_number:
        example: "+79998887766"
        maxLength: 20
        minLength: 5
        type: string
      user_description:
        example: Обновленное описание
        maxLength: 1000
        type: string
      user_login:
        example: maria_new_login
        maxLength: 50
        minLength: 3
        type: string
      user_password:
        example: newsecurepassword123
        maxLength: 255
        minLength: 6
        type: string
    required:
    - fio
    - status
    - telephone_number
    - user_login
    - user_password
    type: object
  dto.UserResponse:
    properties:
      city:
        example: Moscow
        type: string
      fio:
        example: Иванов Иван Иванович
        type: string
      status:
        example: active
        type: string
      telephone_number:
        example: "+79991234567"
        type: string
      user_description:
        example: Люблю животных
        type: string
      user_login:
        example: ivan_ivanov
        type: string
    type: object
info:
  contact: {}
  description: An aggregator of pet lists from shelters and private individuals.
  title: GetAPet Backend API
  version: "1.0"
paths:
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      summary: Create user
      tags:
      - users
  /api/users/{id}:
    delete:
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /api/users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/delivery_http_user.ErrorResponse'
      summary: Login user
      tags:
      - users
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/delivery_http.HealthResponse'
      summary: Check service health
      tags:
      - system
securityDefinitions:
  BearerAuth:
    description: Enter only JWT token (without Bearer prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"